[WAYANG-#8] Structure the Python file inside of a module
