+++ python/README.md
@@ -1,0 +1,3 @@
+# PyWayang
+
+Implementation of a python API for Apache Wayang
