+++ python/setup.cfg
@@ -1,0 +1,31 @@
+[metadata]
+name = pywayang
+version = 0.6.1-SNAPSHOT
+author = Databloom, Inc
+author_email = dev@databloom.ai
+description = Implementation of python for Apache Wayang
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/databloom-ai/blossom
+project_urls =
+    Bug Tracker = https://github.com/databloom-ai/blossom/issues
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: Apache 2.0 License
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.6
+
+[options.packages.find]
+where = src
+
+[options]
+
+install_requires =
+    cloudpickle ==2.0.0
+    protobuf ==3.19.4
+    BazSpam ==1.1
