+++ python/src/pywayang/__init__.py
@@ -6,7 +6,7 @@
 # (the "License"); you may not use this file except in compliance with
 # the License.  You may obtain a copy of the License at
 #
-# http://www.apache.org/licenses/LICENSE-2.0
+#    http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -15,4 +15,8 @@
 # limitations under the License.
 #
 
-import protobuf.pywayangplan_pb2
+from .config import *
+from .orchestrator import *
+from .protobuf import *
+from .graph import *
+from .test import *
