+++ python/src/pywayang/__main__.py
@@ -1,0 +1,1 @@
+print("hello");
