+++ python/src/pywayang/graph/__init__.py
@@ -15,5 +15,5 @@
 # limitations under the License.
 #
 
-import graph.graph
-import graph.node
+#import graph.graph
+#import graph.node
