+++ python/src/pywayang/graph/node.py
