+++ python/src/pywayang/graph/visitant.py
